Change FromVersion field in LoadOptions to int64 for consistency

+++ eventstore.go
@@ -22,7 +22,7 @@ type Event struct {
 // LoadOptions contains options for loading events from a stream.
 type LoadOptions struct {
 	// FromVersion specifies where to start loading events from
-	FromVersion int
+	FromVersion int64
 	// Limit specifies the maximum number of events to return
 	Limit int
 }
